fix: stabilize release ci checks

+++ internal/backups/space_unix.go
@@ -9,5 +9,5 @@ func defaultAvailableBytes(path string) (uint64, bool) {
 	if err := unix.Statfs(path, &stat); err != nil {
 		return 0, false
 	}
-	return uint64(stat.Bavail) * uint64(stat.Bsize), true
+	return stat.Bavail * uint64(stat.Bsize), true
 }
